[cgroups2] ignore manual enabling of perf_event during Cgroups2IsolatorProcess::prepare
In Cgroups2IsolatorProcess::prepare, it may manually enable controller
by writing to the cgroup.subtree_control process.

For perf_event, since is is automatically turned on, it does not appear
inside the cgroup.controllers file and hence cannot be written to the
cgroup.subtree_control file. For this reason, we skip the enable call for
the perf_event controller.

Review: https://reviews.apache.org/r/74998