[http] add support for executor_id and framework_id query parameters in /containers endpoint
We now allow filtering by executor and framework id in addition to the original functionality of filtering by containerId.

Please note that the /containers endooint only allows select combinations of these query parameter fields to be populated at once.
We will return a failure if we see that the combination of query paramters is invalid.

We currently accept:
* no query parameters
* only container id
* only framework id
* only framework id and executor id

Review: https://reviews.apache.org/r/75009