[cgroups2] update device configuration EBPF program generator
In cgroups2, we want our EBPF file to only grant access to a device if it is in a cgroup's allow list and not in its deny list.
This means that we need to change our previous logic that exits on the first match to now check both the allow and deny list of a cgroup
before determining whether access may be granted.

This patch implements the logic change, and removes functions that are no longer necessary for the DeviceProgram class.
We now pass the entire allow and deny list to a configure function inside the DeviceProgram object, which will create a ebpf program
with the updated logic and attempt to attach it to the cgroup.

Review: https://reviews.apache.org/r/75026