[mesos-build] add correct directory to git safe directory & add explanation
Previously, the entrypoint.sh added /SRC/.git as a safe directory after the git clone, which was not useful.
We now add it /SRC/ as a safe directory before the git clone.
In addition, a note has been added to explain why we are adding SRC to the safe directory.

Review: https://reviews.apache.org/r/75029